Move `CCITTFaxStream` and `Jbig2Stream`, from `src/core/stream.js`, to separate files
